quieten defaulting warnings
Ignore-this: 8fdfaef214b42bc2eee9dad9a860f2c5

darcs-hash:1ffa36339bb3db0b639d716e8d004e8a7f0e2b73

+++ Sound/SC3/Lang/Control/Event.hs
@@ -329,7 +329,7 @@ to_sc3_bundle t j e =
              : ("sustain",rt)
              : ("amp",e_amp e)
              : parameters e
-        pr' = map (\(k,v) -> (k,realToFrac v)) pr
+        pr' = map (\(k,v) -> (k,v)) pr
         i = fromMaybe j (e_id e)
         t' = t + realToFrac (latency e)
     in if is_rest e || isNaN f
@@ -341,8 +341,8 @@ to_sc3_bundle t j e =
                 m_off = if not sr
                         then []
                         else case e_type e of
-                               E_s_new -> [S.n_set i [("gate",0)]]
-                               E_n_set -> [S.n_set i [("gate",0)]]
+                               E_s_new -> [S.n_set i [("gate",0::Double)]]
+                               E_n_set -> [S.n_set i [("gate",0::Double)]]
                                E_rest -> []
             in Just (O.Bundle t' m_on
                     ,O.Bundle (t' + realToFrac rt) m_off)
